Change test file names from 'tests' to 'test'
